extract memory buffer from HLS mediasubsession
